arosc.library: Fixes for $(PARANOIA_CFLAGS)
Note, this requires the Exec/Child*() APTR -> ULONG patch as
a prerequisite


git-svn-id: https://svn.aros.org/svn/aros/trunk@37160 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/bcopy.c
@@ -87,13 +87,13 @@ bcopy(const void *src0, void *dst0, size_t length)
 		/*
 		 * Copy forward.
 		 */
-		t = (int)src;	/* only need low bits */
-		if ((t | (int)dst) & wmask) {
+		t = (unsigned long)src;	/* only need low bits */
+		if ((t | (unsigned long)dst) & wmask) {
 			/*
 			 * Try to align operands.  This cannot be done
 			 * unless the low bits match.
 			 */
-			if ((t ^ (int)dst) & wmask || length < wsize)
+			if ((t ^ (unsigned long)dst) & wmask || length < wsize)
 				t = length;
 			else
 				t = wsize - (t & wmask);
@@ -115,9 +115,9 @@ bcopy(const void *src0, void *dst0, size_t length)
 		 */
 		src += length;
 		dst += length;
-		t = (int)src;
-		if ((t | (int)dst) & wmask) {
-			if ((t ^ (int)dst) & wmask || length <= wsize)
+		t = (unsigned long)src;
+		if ((t | (unsigned long)dst) & wmask) {
+			if ((t ^ (unsigned long)dst) & wmask || length <= wsize)
 				t = length;
 			else
 				t &= wmask;
